+++ storage/framework/views/4e06a71a0e1114d8b04c468643af1408f8a83c17.php
@@ -33,7 +33,7 @@
                 <td><span class="badge bg-success text-white"><?php echo e($j->nama_kategori); ?></span></td>
                 <td>
                     <a href="/jasa/detail/<?php echo e($j->id); ?>" class="btn btn-primary btn-sm"><i class="fas fa-eye"></i> Detail</a>
-                    <a href="/jasa/hapus/<?php echo e($j->id); ?>" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i> Hapus</a>
+                    <a href="/jasa/hapus/<?php echo e($j->id); ?>" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt" onclick="javascript:return confirm('apakah yakin akan menghapus data ?')"></i> Hapus</a>
                 </td>
             </tr>
             <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?>
